+++ contrib/sacrebleu/sacrebleu.py
@@ -84,6 +84,17 @@ Sacre BLEU.
 
 # VERSION HISTORY
 
+- 1.2.6 (22 March 2018)
+   - added wmt17/ms (Microsoft's [additional ZH-EN references](https://github.com/MicrosoftTranslator/Translator-HumanParityData)).
+     Try `sacrebleu -t wmt17/ms --cite`.
+   - `--echo ref` now pastes together all references, if there is more than one
+
+- 1.2.5 (13 March 2018)
+   - added wmt18/dev datasets (en-et and et-en)
+   - fixed logic with --force
+   - locale-independent installation
+   - added "--echo both" (tab-delimited)
+
 - 1.2.3 (28 January 2018)
    - metrics (`-m`) are now printed in the order requested
    - chrF now prints a version string (including the beta parameter, importantly)
@@ -161,7 +172,6 @@ import io
 import os
 import re
 import sys
-import tarfile
 import urllib.request
 from collections import Counter, namedtuple
 from itertools import zip_longest
@@ -170,7 +180,7 @@ from typing import List, Iterable, Tuple
 import math
 import unicodedata
 
-VERSION = '1.2.3'
+VERSION = '1.2.6'
 
 try:
     # SIGPIPE is not available on Windows machines, throwing an exception.
@@ -207,6 +217,12 @@ CHRF_BETA = 2
 # Many of these are *.sgm files, which are processed to produced plain text that can be used by this script.
 # The canonical location of unpacked, processed data is $SACREBLEU/$TEST/$SOURCE-$TARGET.{$SOURCE,$TARGET}
 DATASETS = {
+    'wmt18/dev': {
+        'data': ['http://data.statmt.org/wmt18/translation-task/dev.tgz'],
+        'description': 'Development data (Estonian<>English).',
+        'et-en': ['dev/newsdev2018-eten-src.et.sgm', 'dev/newsdev2018-eten-ref.en.sgm'],
+        'en-et': ['dev/newsdev2018-enet-src.en.sgm', 'dev/newsdev2018-enet-ref.et.sgm'],
+    },
     'wmt17': {
         'data': ['http://data.statmt.org/wmt17/translation-task/test.tgz'],
         'description': 'Official evaluation data.',
@@ -249,6 +265,13 @@ DATASETS = {
         'en-zh': ['dev/newsdev2017-enzh-src.en.sgm', 'dev/newsdev2017-enzh-ref.zh.sgm'],
         'lv-en': ['dev/newsdev2017-lven-src.lv.sgm', 'dev/newsdev2017-lven-ref.en.sgm'],
         'zh-en': ['dev/newsdev2017-zhen-src.zh.sgm', 'dev/newsdev2017-zhen-ref.en.sgm'],
+    },
+    'wmt17/ms': {
+        'data': ['https://github.com/MicrosoftTranslator/Translator-HumanParityData/archive/master.zip',
+                 'http://data.statmt.org/wmt17/translation-task/test-update-1.tgz'],
+        'description': 'Additional Chinese-English references from Microsoft Research.',
+        'citation': '@inproceedings{achieving-human-parity-on-automatic-chinese-to-english-news-translation,\n  author = {Hassan Awadalla, Hany and Aue, Anthony and Chen, Chang and Chowdhary, Vishal and Clark, Jonathan and Federmann, Christian and Huang, Xuedong and Junczys-Dowmunt, Marcin and Lewis, Will and Li, Mu and Liu, Shujie and Liu, Tie-Yan and Luo, Renqian and Menezes, Arul and Qin, Tao and Seide, Frank and Tan, Xu and Tian, Fei and Wu, Lijun and Wu, Shuangzhi and Xia, Yingce and Zhang, Dongdong and Zhang, Zhirui and Zhou, Ming},\n  title = {Achieving Human Parity on Automatic Chinese to English News Translation},\n  booktitle = {},\n  year = {2018},\n  month = {March},\n  abstract = {Machine translation has made rapid advances in recent years. Millions of people are using it today in online translation systems and mobile applications in order to communicate across language barriers. The question naturally arises whether such systems can approach or achieve parity with human translations. In this paper, we first address the problem of how to define and accurately measure human parity in translation. We then describe Microsoft’s machine translation system and measure the quality of its translations on the widely used WMT 2017 news translation task from Chinese to English. We find that our latest neural machine translation system has reached a new state-of-the-art, and that the translation quality is at human parity when compared to professional human translations. We also find that it significantly exceeds the quality of crowd-sourced non-professional translations.},\n  publisher = {},\n  url = {https://www.microsoft.com/en-us/research/publication/achieving-human-parity-on-automatic-chinese-to-english-news-translation/},\n  address = {},\n  pages = {},\n  journal = {},\n  volume = {},\n  chapter = {},\n  isbn = {},\n}',
+        'zh-en': ['newstest2017-zhen-src.zh.sgm', 'newstest2017-zhen-ref.en.sgm', 'Translator-HumanParityData-master/Translator-HumanParityData/References/Translator-HumanParityData-Reference-HT.txt', 'Translator-HumanParityData-master/Translator-HumanParityData/References/Translator-HumanParityData-Reference-PE.txt'],
     },
     'wmt16': {
         'data': ['http://data.statmt.org/wmt16/translation-task/test.tgz'],
@@ -445,7 +468,7 @@ DATASETS = {
                  'https://wit3.fbk.eu/archive/2017-01-ted-test/texts/en/zh/en-zh.tgz',
                  'https://wit3.fbk.eu/archive/2017-01-ted-test/texts/zh/en/zh-en.tgz'],
         'description': 'Official evaluation data for IWSLT.',
-        'citation': '@InProceedings{iwslt2017,\n  author    = {Cettolo, Mauro and Federico, Marcello and Bentivogli, Luisa and Niehues, Jan and Stüker, Sebastian and Sudoh, Katsuitho and Yoshino, Koichiro and Federmann, Christian},\n  title     = {Overview of the IWSLT 2017 Evaluation Campaign},\n  booktitle = {14th International Workshop on Spoken Language Translation},\n  month     = {December},\n  year      = {2017},\n  address   = {Tokyo, Japan},\n  pages     = {2--14},\n  url       = {http://workshop2017.iwslt.org/downloads/iwslt2017_proceeding_v2.pdf\n}',
+        'citation': '@InProceedings{iwslt2017,\n  author    = {Cettolo, Mauro and Federico, Marcello and Bentivogli, Luisa and Niehues, Jan and Stüker, Sebastian and Sudoh, Katsuitho and Yoshino, Koichiro and Federmann, Christian},\n  title     = {Overview of the IWSLT 2017 Evaluation Campaign},\n  booktitle = {14th International Workshop on Spoken Language Translation},\n  month     = {December},\n  year      = {2017},\n  address   = {Tokyo, Japan},\n  pages     = {2--14},\n  url       = {http://workshop2017.iwslt.org/downloads/iwslt2017_proceeding_v2.pdf}\n}',
         'en-fr': ['en-fr/IWSLT17.TED.tst2017.en-fr.en.xml', 'fr-en/IWSLT17.TED.tst2017.fr-en.fr.xml'],
         'fr-en': ['fr-en/IWSLT17.TED.tst2017.fr-en.fr.xml', 'en-fr/IWSLT17.TED.tst2017.en-fr.en.xml'],
         'en-de': ['en-de/IWSLT17.TED.tst2017.en-de.en.xml', 'de-en/IWSLT17.TED.tst2017.de-en.de.xml'],
@@ -774,14 +797,14 @@ TOKENIZERS = {
 DEFAULT_TOKENIZER = '13a'
 
 
-def _open(file, encoding='utf-8'):
+def smart_open(file, mode='rt', encoding='utf-8'):
     """Convenience function for reading compressed or plain text files.
     :param file: The file to read.
     :param encoding: The file encoding.
     """
     if file.endswith('.gz'):
-        return gzip.open(file, 'rt', encoding=encoding)
-    return open(file, 'rt', encoding=encoding)
+        return gzip.open(file, mode=mode, encoding=encoding)
+    return open(file, mode=mode, encoding=encoding)
 
 
 def my_log(num):
@@ -934,15 +957,19 @@ def process_to_text(rawfile, txtfile):
     if not os.path.exists(txtfile) or os.path.getsize(txtfile) == 0:
         logging.info("Processing %s to %s", rawfile, txtfile)
         if rawfile.endswith('.sgm') or rawfile.endswith('.sgml'):
-            with _open(rawfile) as fin, open(txtfile, 'wt') as fout:
+            with smart_open(rawfile) as fin, smart_open(txtfile, 'wt') as fout:
                 for line in fin:
                     if line.startswith('<seg '):
                         print(_clean(re.sub(r'<seg.*?>(.*)</seg>.*?', '\\1', line)), file=fout)
         elif rawfile.endswith('.xml'): # IWSLT
-            with _open(rawfile) as fin, open(txtfile, 'wt') as fout:
+            with smart_open(rawfile) as fin, smart_open(txtfile, 'wt') as fout:
                 for line in fin:
                     if line.startswith('<seg '):
                         print(_clean(re.sub(r'<seg.*?>(.*)</seg>.*?', '\\1', line)), file=fout)
+        elif rawfile.endswith('.txt'): # wmt17/ms
+            with smart_open(rawfile) as fin, smart_open(txtfile, 'wt') as fout:
+                for line in fin:
+                    print(line.rstrip(), file=fout)
 
 def print_test_set(test_set, langpair, side):
     """Prints to STDOUT the specified side of the specified test set
@@ -951,11 +978,15 @@ def print_test_set(test_set, langpair, side):
     :param side: 'src' for source, 'ref' for reference
     """
 
-    where = download_test_set(test_set, langpair)
-    infile = where[0] if side == 'src' else where[1]
-    with open(infile) as fin:
-        for line in fin:
-            print(line.rstrip())
+    files = download_test_set(test_set, langpair)
+    if side == 'src':
+        files = [files[0]]
+    elif side == 'ref':
+        files.pop(0)
+
+    streams = [smart_open(file) for file in files]
+    for lines in zip(*streams):
+        print('\t'.join(map(lambda x: x.rstrip(), lines)))
 
 
 def download_test_set(test_set, langpair=None):
@@ -964,9 +995,6 @@ def download_test_set(test_set, langpair=None):
     :param langpair: the language pair (needed for some datasets)
     :return: the set of processed files
     """
-
-    # if not data.has_key(test_set):
-    #     return None
 
     outdir = os.path.join(SACREBLEU, test_set)
     if not os.path.exists(outdir):
@@ -990,8 +1018,15 @@ def download_test_set(test_set, langpair=None):
 
             # Extract the tarball
             logging.info('Extracting %s', tarball)
-            tar = tarfile.open(tarball)
-            tar.extractall(path=rawdir)
+            if tarball.endswith('.tar.gz') or tarball.endswith('.tgz'):
+                import tarfile
+                tar = tarfile.open(tarball)
+                tar.extractall(path=rawdir)
+            elif tarball.endswith('.zip'):
+                import zipfile
+                zipfile = zipfile.ZipFile(tarball, 'r')
+                zipfile.extractall(path=rawdir)
+                zipfile.close()
 
     found = []
 
@@ -1106,13 +1141,13 @@ def corpus_bleu(sys_stream, ref_streams, smooth='exp', smooth_floor=0.0, force=F
         if lowercase:
             lines = [x.lower() for x in lines]
 
-        if (not force or tokenize != 'none') and lines[0].rstrip().endswith(' .'):
+        if not (force or tokenize == 'none') and lines[0].rstrip().endswith(' .'):
             tokenized_count += 1
 
             if tokenized_count == 100:
-                logging.warning('That\'s > 100 lines that end in a tokenized period (\'.\')')
+                logging.warning('That\'s 100 lines that end in a tokenized period (\'.\')')
                 logging.warning('It looks like you forgot to detokenize your test data, which may hurt your score.')
-                logging.warning('If you insist your data is tokenized, you can suppress this message with \'--force\'.')
+                logging.warning('If you insist your data is detokenized, or don\'t care, you can suppress this message with \'--force\'.')
 
         output, *refs = [TOKENIZERS[tokenize](x.rstrip()) for x in lines]
 
@@ -1263,8 +1298,8 @@ def main():
                             help='source-target language pair (2-char ISO639-1 codes)')
     arg_parser.add_argument('--download', type=str, default=None,
                             help='download a test set and quit')
-    arg_parser.add_argument('--echo', choices=['src', 'ref'], type=str, default=None,
-                            help='output the source or reference to STDOUT and quit')
+    arg_parser.add_argument('--echo', choices=['src', 'ref', 'both'], type=str, default=None,
+                            help='output the source (src), reference (ref), or both (both, pasted) to STDOUT and quit')
     arg_parser.add_argument('--input', '-i', type=str, default='-',
                             help='Read input from a file instead of STDIN')
     arg_parser.add_argument('refs', nargs='*', default=[],
@@ -1348,11 +1383,11 @@ def main():
     else:
         refs = args.refs
 
-    inputfh = io.TextIOWrapper(sys.stdin.buffer, encoding=args.encoding) if args.input == '-' else _open(args.input, args.encoding)
+    inputfh = io.TextIOWrapper(sys.stdin.buffer, encoding=args.encoding) if args.input == '-' else smart_open(args.input, encoding=args.encoding)
     system = inputfh.readlines()
 
     # Read references
-    refs = [_open(x, args.encoding).readlines() for x in refs]
+    refs = [smart_open(x, encoding=args.encoding).readlines() for x in refs]
 
     if args.langpair is not None:
         _, target = args.langpair.split('-')