+++ sockeye/training.py
@@ -53,6 +53,7 @@ class TrainingModel(model.SockeyeModel):
     :param bucketing: If True bucketing will be used, if False the computation graph will always be
             unrolled to the full length.
     :param gradient_compression_params: Optional dictionary of gradient compression parameters.
+    :param fixed_param_names: Optional list of params to fix during training (i.e. their values will not be trained).
     """
 
     def __init__(self,
@@ -63,10 +64,12 @@ class TrainingModel(model.SockeyeModel):
                  provide_label: List[mx.io.DataDesc],
                  default_bucket_key: Tuple[int, int],
                  bucketing: bool,
-                 gradient_compression_params: Optional[Dict[str, Any]] = None) -> None:
+                 gradient_compression_params: Optional[Dict[str, Any]] = None,
+                 fixed_param_names: Optional[List[str]] = None) -> None:
         super().__init__(config)
         self.context = context
         self.output_dir = output_dir
+        self.fixed_param_names = fixed_param_names
         self._bucketing = bucketing
         self._gradient_compression_params = gradient_compression_params
         self._initialize(provide_data, provide_label, default_bucket_key)
@@ -164,7 +167,8 @@ class TrainingModel(model.SockeyeModel):
                                                  logger=logger,
                                                  default_bucket_key=default_bucket_key,
                                                  context=self.context,
-                                                 compression_params=self._gradient_compression_params)
+                                                 compression_params=self._gradient_compression_params,
+                                                 fixed_param_names=self.fixed_param_names)
         else:
             logger.info("No bucketing. Unrolled to (%d,%d)",
                         self.config.config_data.max_seq_len_source, self.config.config_data.max_seq_len_target)
@@ -174,7 +178,8 @@ class TrainingModel(model.SockeyeModel):
                                         label_names=label_names,
                                         logger=logger,
                                         context=self.context,
-                                        compression_params=self._gradient_compression_params)
+                                        compression_params=self._gradient_compression_params,
+                                        fixed_param_names=self.fixed_param_names)
 
         self.module.bind(data_shapes=provide_data,
                          label_shapes=provide_label,
@@ -307,6 +312,8 @@ class TrainingModel(model.SockeyeModel):
             info.append("%s: %s" % (name, array.shape))
             total_parameters += reduce(lambda x, y: x * y, array.shape)
         logger.info("Model parameters: %s", ", ".join(info))
+        if self.fixed_param_names:
+            logger.info("Fixed model parameters: %s", ", ".join(self.fixed_param_names))
         logger.info("Total # of parameters: %d", total_parameters)
 
     def save_params_to_file(self, fname: str):