+++ sockeye/utils.py
@@ -258,9 +258,7 @@ def smallest_k(matrix: np.ndarray, k: int,
         flatten = matrix.flatten()
 
     # args are the indices in flatten of the k smallest elements
-    args = np.argpartition(flatten, k)[:k]
-    # args are the indices in flatten of the sorted k smallest elements
-    args = args[np.argsort(flatten[args])]
+    args = np.argpartition(flatten, range(k))[:k]
     # flatten[args] are the values for args
     return np.unravel_index(args, matrix.shape), flatten[args]